attempt to open JSON (not succesful)
