+++ Contexts/Contexts.ts
@@ -1,5 +1,5 @@
 import React from 'react';
-import { IGutkaCtx, IGlobalCtx, IViewerCtx } from '../Config/interfaces';
+import { IGutkaCtx, IGlobalCtx, IViewerCtx } from '../config/interfaces';
 
 export const GutkaContext = React.createContext<IGutkaCtx>({
   gutkas: [],
