+++ package.json
@@ -18,6 +18,7 @@
     "react-native": "0.61.5",
     "react-native-async-storage": "0.0.1",
     "react-native-elements": "^1.2.7",
+    "react-native-fs": "^2.16.4",
     "react-native-gesture-handler": "^1.5.2",
     "react-native-reanimated": "^1.4.0",
     "react-native-vector-icons": "^6.6.0",
